v15

+++ src/main/java/ru/kata/spring/boot_security/demo/DAO/UserDAO.java
@@ -1,13 +1,14 @@
 package ru.kata.spring.boot_security.demo.DAO;
 
+import org.springframework.security.core.userdetails.UserDetailsService;
 import ru.kata.spring.boot_security.demo.model.User;
 import java.util.List;
 
 public interface UserDAO {
     List<User> getAllUsers();
-    User getUserById(int id);
-    void updateUser(int id, User user);
-    void deleteById(int id);
+    User getUserById(Long id);
+    void updateUser(Long id, User user);
+    void deleteById(Long id);
     void addUser(User user);
     User getUserByName(String name);
 }
